Refactor the topology system
The Topology type has been removed, the Universe directly owns the
list of Molecules in the system

The Molecule type now contains all the information about the bonding in
the system.

Molecules are continguous in memory in order to facilitate future
MonteCarlo implementation, and iteration over the particles in a
molecule. The add_bond and related functions are thus more complex,
to handle the possible movements of particles when a new bond force
two molecules to be merged.